Extract common Webpack settings into a base configuration
