+++ package.json
@@ -80,6 +80,7 @@
     "superagent": "^1.8.3",
     "webpack": "^1.12.14",
     "webpack-dev-server": "^1.14.1",
+    "webpack-merge": "^0.12.0",
     "wpcom": "^4.9.1",
     "wpcom-proxy-request": "^2.0.0",
     "wpcom-unpublished": "^1.1.0"
